backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
loguru==0.7.2
multidict==6.1.0
passlib==1.7.4
playwright==1.48.0
propcache==0.2.0
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==12.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
watchgod==0.8.2
win32-setctime==1.1.0
yarl==1.17.1
